requirements.txt
astroid==3.3.9
certifi==2025.1.31
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.56.0
hmmlearn==0.3.3
isort==6.0.1
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pulsar-client==3.6.1
pylint==3.3.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.5.0
tomlkit==0.13.2
